Consolidate packages
