+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/profiler/ProfilerServiceTest.java
@@ -25,9 +25,9 @@ import com.microsoft.applicationinsights.profiler.ProfilerServiceFactory;
 import com.microsoft.applicationinsights.profiler.config.ServiceProfilerServiceConfig;
 import com.microsoft.applicationinsights.profiler.uploader.UploadCompleteHandler;
 import com.microsoft.applicationinsights.serviceprofilerapi.JfrProfilerService;
+import com.microsoft.applicationinsights.serviceprofilerapi.client.ArtifactAcceptedResponse;
+import com.microsoft.applicationinsights.serviceprofilerapi.client.BlobAccessPass;
 import com.microsoft.applicationinsights.serviceprofilerapi.client.ServiceProfilerClient;
-import com.microsoft.applicationinsights.serviceprofilerapi.client.contract.ArtifactAcceptedResponse;
-import com.microsoft.applicationinsights.serviceprofilerapi.client.contract.BlobAccessPass;
 import com.microsoft.applicationinsights.serviceprofilerapi.profiler.JfrProfiler;
 import com.microsoft.applicationinsights.serviceprofilerapi.upload.ServiceProfilerUploader;
 import com.microsoft.applicationinsights.serviceprofilerapi.upload.UploadContext;
